+++ test/base/src/androidTest/kotlin/io/realm/test/shared/MutableRealmTests.kt
@@ -15,10 +15,13 @@
  */
 package io.realm.test.shared
 
+import io.realm.MutableRealm
 import io.realm.Realm
 import io.realm.RealmConfiguration
+import io.realm.RealmInstant
 import io.realm.RealmResults
 import io.realm.entities.Sample
+import io.realm.entities.SampleWithPrimaryKey
 import io.realm.entities.StringPropertyWithPrimaryKey
 import io.realm.entities.link.Child
 import io.realm.entities.link.Parent
@@ -30,6 +33,7 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
+import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -37,6 +41,7 @@ import kotlin.test.assertNotEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
+import kotlin.test.fail
 
 class MutableRealmTests {
 
@@ -52,7 +57,8 @@ class MutableRealmTests {
                 Parent::class,
                 Child::class,
                 StringPropertyWithPrimaryKey::class,
-                Sample::class
+                Sample::class,
+                SampleWithPrimaryKey::class
             )
         ).path("$tmpDir/default.realm").build()
         realm = Realm.open(configuration)
@@ -67,11 +73,291 @@ class MutableRealmTests {
     }
 
     @Test
-    fun copyToRealmWithDefaults() {
+    fun copyToRealm_updatePolicy_error_withDefaults() {
         realm.writeBlocking { copyToRealm(Parent()) }
         val parents = realm.query<Parent>().find()
         assertEquals(1, parents.size)
         assertEquals("N.N.", parents[0].name)
+    }
+
+    @Test
+    fun copyToRealm_updatePolicy_error_throwsOnDuplicatePrimaryKey() {
+        realm.writeBlocking {
+            copyToRealm(SampleWithPrimaryKey())
+            assertFailsWith<IllegalArgumentException> {
+                copyToRealm(SampleWithPrimaryKey())
+            }
+        }
+        assertEquals(1, realm.query<SampleWithPrimaryKey>().find().size)
+    }
+
+    @Test
+    fun set_throwsOnDuplicatePrimaryKey() {
+        realm.writeBlocking {
+            val sample = copyToRealm(SampleWithPrimaryKey())
+            assertFailsWith<IllegalArgumentException> {
+                sample.nullableObject = SampleWithPrimaryKey()
+            }
+        }
+        assertEquals(1, realm.query<SampleWithPrimaryKey>().find().size)
+    }
+
+    @Test
+    fun copyToRealm_updatePolicy_all() {
+        realm.writeBlocking {
+            val obj = StringPropertyWithPrimaryKey()
+            copyToRealm(obj.apply { id = "PRIMARY_KEY" })
+
+            obj.apply { value = "UPDATED_VALUE" }
+            copyToRealm(obj, MutableRealm.UpdatePolicy.ALL)
+        }
+
+        val objects = realm.query<StringPropertyWithPrimaryKey>().find()
+        assertEquals(1, objects.size)
+        objects[0].run {
+            assertEquals("PRIMARY_KEY", id)
+            assertEquals("UPDATED_VALUE", value)
+        }
+    }
+
+    @Test
+    fun copyToRealm_updatePolicy_all_nonPrimaryKeyField() {
+        realm.writeBlocking {
+            copyToRealm(Parent(), MutableRealm.UpdatePolicy.ALL)
+        }
+        assertEquals(1, realm.query<Parent>().find().size)
+    }
+
+    @Test
+    @Suppress("LongMethod")
+    fun copyToRealm_updatePolicy_all_allTypes() {
+        realm.writeBlocking {
+            copyToRealm(
+                SampleWithPrimaryKey().apply {
+                    primaryKey = 1
+                    stringField = "ORIGINAL"
+                }
+            )
+        }
+        assertEquals(1, realm.query<SampleWithPrimaryKey>().count().find())
+
+        // TODO Verify that we cover all types
+        val sample = SampleWithPrimaryKey().apply {
+            primaryKey = 1
+            stringField = "UPDATED"
+            byteField = 0x10
+            charField = 'b'
+            shortField = 255
+            intField = 255
+            longField = 1024
+            booleanField = false
+            floatField = 42.42f
+            doubleField = 42.42
+            timestampField = RealmInstant.fromEpochSeconds(42, 42)
+
+            nullableStringField = "UPDATED"
+            nullableByteField = 0x10
+            nullableCharField = 'b'
+            nullableShortField = 255
+            nullableIntField = 255
+            nullableLongField = 1024
+            nullableBooleanField = false
+            nullableFloatField = 42.42f
+            nullableDoubleField = 42.42
+            nullableTimestampField = RealmInstant.fromEpochSeconds(42, 42)
+            nullableObject = this
+
+            stringListField.add("UPDATED")
+            byteListField.add(0x10)
+            charListField.add('b')
+            shortListField.add(255)
+            intListField.add(255)
+            longListField.add(1024)
+            booleanListField.add(false)
+            floatListField.add(3.14f)
+            doubleListField.add(3.14)
+            timestampListField.add(RealmInstant.fromEpochSeconds(42, 42))
+
+            objectListField.add(this)
+            nullableStringListField.add(null)
+            nullableByteListField.add(null)
+            nullableCharListField.add(null)
+            nullableShortListField.add(null)
+            nullableIntListField.add(null)
+            nullableLongListField.add(null)
+            nullableBooleanListField.add(null)
+            nullableFloatListField.add(null)
+            nullableDoubleListField.add(null)
+            nullableTimestampListField.add(null)
+        }
+        realm.writeBlocking { copyToRealm(sample, MutableRealm.UpdatePolicy.ALL) }
+
+        val samples = realm.query<SampleWithPrimaryKey>().find()
+        assertEquals(1, samples.size)
+        samples[0].run {
+            assertEquals("UPDATED", stringField)
+            assertEquals(0x10, byteField)
+            assertEquals('b', charField)
+            assertEquals(255, shortField)
+            assertEquals(255, intField)
+            assertEquals(1024, longField)
+            assertEquals(false, booleanField)
+            assertEquals(42.42f, floatField)
+            assertEquals(42.42, doubleField)
+            assertEquals(RealmInstant.fromEpochSeconds(42, 42), timestampField)
+
+            assertEquals("UPDATED", nullableStringField)
+            assertEquals(0x10, nullableByteField)
+            assertEquals('b', nullableCharField)
+            assertEquals(255, nullableShortField)
+            assertEquals(255, nullableIntField)
+            assertEquals(1024, nullableLongField)
+            assertEquals(false, nullableBooleanField)
+            assertEquals(42.42f, nullableFloatField)
+            assertEquals(42.42, nullableDoubleField)
+            assertEquals(RealmInstant.fromEpochSeconds(42, 42), nullableTimestampField)
+            assertEquals(primaryKey, nullableObject!!.primaryKey)
+
+            assertEquals("UPDATED", stringListField[0])
+            assertEquals(0x10, byteListField[0])
+            assertEquals('b', charListField[0])
+            assertEquals(255, shortListField[0])
+            assertEquals(255, intListField[0])
+            assertEquals(1024, longListField[0])
+            assertEquals(false, booleanListField[0])
+            assertEquals(3.14f, floatListField[0])
+            assertEquals(3.14, doubleListField[0])
+            assertEquals(RealmInstant.fromEpochSeconds(42, 42), timestampListField[0])
+            assertEquals(primaryKey, objectListField[0].primaryKey)
+
+            assertEquals(null, nullableStringListField[0])
+            assertEquals(null, nullableByteListField[0])
+            assertEquals(null, nullableCharListField[0])
+            assertEquals(null, nullableShortListField[0])
+            assertEquals(null, nullableIntListField[0])
+            assertEquals(null, nullableLongListField[0])
+            assertEquals(null, nullableBooleanListField[0])
+            assertEquals(null, nullableFloatListField[0])
+            assertEquals(null, nullableDoubleListField[0])
+            assertEquals(null, nullableTimestampListField[0])
+        }
+    }
+
+    @Test
+    fun copyToRealm_updatePolicy_all_cyclicObject() {
+        val sample11 = SampleWithPrimaryKey().apply {
+            primaryKey = 1
+            stringField = "One"
+        }
+        val sample22 = SampleWithPrimaryKey().apply {
+            primaryKey = 2
+            stringField = "Two"
+        }
+        sample11.nullableObject = sample22
+        sample22.nullableObject = sample11
+
+        realm.writeBlocking {
+            copyToRealm(sample11)
+        }.run {
+            assertEquals(1, primaryKey)
+            assertEquals("One", stringField)
+            nullableObject?.run {
+                assertEquals(2, primaryKey)
+                assertEquals("Two", stringField)
+            } ?: fail("Object shouldn't be null")
+        }
+
+        // We need to replicate objects as we cannot update them after passing it to another thread
+        // on Kotlin Native
+        val sample13 = SampleWithPrimaryKey().apply {
+            primaryKey = 1
+            stringField = "Three"
+        }
+        val sample24 = SampleWithPrimaryKey().apply {
+            primaryKey = 2
+            stringField = "Four"
+        }
+        sample13.nullableObject = sample24
+        sample24.nullableObject = sample13
+
+        realm.writeBlocking {
+            copyToRealm(sample13, MutableRealm.UpdatePolicy.ALL)
+        }.run {
+            assertEquals(1, primaryKey)
+            assertEquals("Three", stringField)
+            nullableObject?.run {
+                assertEquals(2, primaryKey)
+                assertEquals("Four", stringField)
+            } ?: fail("Object shouldn't be null")
+        }
+    }
+
+    // TODO The cache maintained during import doesn't recognize previously imported object
+    @Ignore // https://github.com/realm/realm-kotlin/issues/708
+    @Test
+    fun copyToRealm_updatePolicy_all_realmJavaBug4957() {
+        val parent = SampleWithPrimaryKey().apply {
+            primaryKey = 0
+
+            val listElement = SampleWithPrimaryKey().apply { primaryKey = 1 }
+            objectListField.add(listElement)
+
+            nullableObject = SampleWithPrimaryKey().apply {
+                primaryKey = 0
+                objectListField.add(listElement)
+                nullableObject = this
+            }
+        }
+        realm.writeBlocking {
+            copyToRealm(parent, MutableRealm.UpdatePolicy.ALL)
+        }.run {
+            assertEquals(1, objectListField.size)
+        }
+    }
+
+    @Test
+    fun copyToRealm_throwsOnManagedObjectFromDifferentVersion() {
+        val frozenParent = realm.writeBlocking { copyToRealm(Parent()) }
+
+        realm.writeBlocking {
+            assertFailsWith<IllegalArgumentException> {
+                copyToRealm(frozenParent)
+            }
+        }
+    }
+
+    @Test
+    fun copyToRealm_throwsWithDeletedObject() {
+        realm.writeBlocking {
+            val parent = copyToRealm(Parent())
+            delete(parent)
+            assertFailsWith<IllegalArgumentException> {
+                copyToRealm(parent)
+            }
+        }
+    }
+
+    @Test
+    fun set_throwsOnManagedObjectFromDifferentVersion() {
+        val child = realm.writeBlocking { copyToRealm(Child()).apply { name = "CHILD" } }
+        realm.writeBlocking {
+            val parent = copyToRealm(Parent())
+            assertFailsWith<IllegalArgumentException> {
+                parent.child = child
+            }
+        }
+    }
+
+    @Test
+    fun set_throwWithDeletedObject() {
+        val child = realm.writeBlocking { copyToRealm(Child()).apply { name = "CHILD" } }
+        realm.writeBlocking {
+            findLatest(child)?.let { delete(it) }
+            val parent = copyToRealm(Parent())
+            assertFailsWith<IllegalArgumentException> {
+                parent.child = child
+            }
+        }
     }
 
     @Test